[Gold II] Title: K번째 수, Time: 148 ms, Memory: 12896 KB -BaekjoonHub

+++ 백준/Gold/1300. K번째 수/K번째 수.java
@@ -3,9 +3,8 @@ import java.util.Scanner;
 public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		long N = sc.nextInt();
-		long K = sc.nextInt();
-
+		long N = sc.nextLong();
+		long K = sc.nextLong();
 		long left = 1;
 		long right = K;
 		long ans = 0;
